Replace 'long' with 'BigInteger' for Business Context services.

+++ src/main/java/org/oagi/srt/gateway/http/api/bie_management/data/BieCreateResponse.java
@@ -2,9 +2,11 @@ package org.oagi.srt.gateway.http.api.bie_management.data;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class BieCreateResponse {
 
-    private long topLevelAbieId;
+    private BigInteger topLevelAbieId;
 
 }
