+++ src/main/java/org/oagi/srt/gateway/http/api/context_management/data/DeleteBusinessContextRequest.java
@@ -2,11 +2,12 @@ package org.oagi.srt.gateway.http.api.context_management.data;
 
 import lombok.Data;
 
+import java.math.BigInteger;
 import java.util.Collections;
 import java.util.List;
 
 @Data
 public class DeleteBusinessContextRequest {
 
-    private List<Long> bizCtxIds = Collections.emptyList();
+    private List<BigInteger> bizCtxIds = Collections.emptyList();
 }
