+++ src/test/java/org/oagi/srt/gateway/http/api/bie_management/controller/BieCreateControllerTest.java
@@ -97,6 +97,6 @@ public class BieCreateControllerTest {
                 );
 
         assertNotNull(response);
-        assertTrue(response.getTopLevelAbieId() > 0L);
+        assertTrue(response.getTopLevelAbieId().longValue() > 0L);
     }
 }
